requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
drf-yasg[validation]~=1.20.0
django-environ~=0.8.1
psycopg2-binary~=2.9.3
progressbar2==4.0.0
coverage==6.4.4